+++ changelog.md
@@ -6,6 +6,8 @@ and this project adheres to http://semver.org/spec/v2.0.0.html.
 
 ## [unreleased]
 
+- SetApplicable understands id values as patterns if given
+- Removed SetApplicableBy, use SetApplicable
 - SetManuallyVerified fails if entry is not applicable
 
 ## [0.8.0] - 2021-03-05